Remove final task that waits for freshclam to run
freshclam will run at next boot on Debian and Ubuntu, since we have
enabled the service.  In the case of Fedora, we just ran freshclam and
verified the presence of /var/lib/clamav/bytecode.cvd.

A further issue is that my TravisCI builds keep petering out on this
task, even though it runs without a problem on my local machine.  I
think this is the age-old problem we have seen with Travis in the
past; too much network activity stalls the build.