Only disable AppArmor profile if AppArmor is enabled in the kernel
Otherwise we would get an error when running molecule tests on a host
that does not have AppArmor enabled.

Also ignore errors when running the aa-disable comment.  This is
suboptimal, but necessary.  Note that disabling the profile means that
it is removed from the kernel.  When we are testing in molecule we
have several Docker containers on the same host sharing the same
kernel, so the first Docker container to disable the profile will
disable it for all the other Docker containers.  This task will then
return an error when the other Docker containers (which will not have
the /etc/apparmor.d/disable/usr.bin.freshclam file) run.