Move AppArmor code to prepare.yml playbook
This way we can add our workaround for the AppArmor+Docker+freshclam
bug only for molecule testing.  The prepare.yml playbook is only
executed by molecule, so we are not altering the ClamAV Ansible role
itself.  This is much cleaner.

Also switch from aa-disable to aa-complain.  In complain mode the
AppArmor policy violations are allowed but logged, so there is still a
record of them.  Furthermore, you can run aa-complain multiple times
without an error, which means we no longer need to ignore errors from
the aa-* step.