Remove some ARM64 platforms from the build workflow
- Debian 13 has a very recent version of systemd (currently 257), and
the freshclam service cannot be started under QEMU emulation.  We
support this platform, but we cannot test it until we have native
ARM64 runners.  See #82 for more details.
- Kali has a recent version of systemd (currently 256), and the
freshclam service cannot be started under QEMU emulation.  We support
this platform, but we cannot test it until we have native ARM64
runners.  See #83 for more details.